+++ pkg/server/addMember.go
@@ -8,7 +8,7 @@ import (
 func (s *Server) AddMember(ctx context.Context, req *proto.AddMemberRequest) (*proto.AddMemberResponse, error) {
 
 	for _, u := range req.MembersId {
-		err := s.db.CreateChatMember(uint(req.ChatId), u, "member")
+		err := s.db.CreateChatMember(req.ChatId, u, "member")
 		if err != nil {
 			return nil, err
 		}
